CLOUD/drawsguard-api-collector-fixed/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
google-cloud-bigquery==3.11.4
google-cloud-secret-manager==2.16.4
google-cloud-logging==3.5.0
requests==2.31.0
pytz==2023.3
aiohttp==3.9.1

